fix: useSearchParams 대신 브라우저 API 사용
